// demos/ui-extended/src/components/schema-builder/app/types.ts
export interface CardComponentPropsType {
  name: string;
  required?: boolean;
  hideKey?: boolean;
  definitionData?: { [key: string]: any };
  definitionUi?: { [key: string]: any };
  neighborNames?: string[];
  dependents?: { children: string[]; value?: any }[];
  dependent?: boolean;
  parent?: string;
  'ui:options'?: { [key: string]: any };
  category?: string;
  schema?: { [key: string]: any };
  type?: string;
  'ui:column'?: string;
  minLength?: number;
  maxLength?: number;
  pattern?: string;
  'ui:autofocus'?: boolean;
  'ui:placeholder'?: string;
  minItems?: number;
  maxItems?: number;
  title?: string;
  $ref?: string;
  format?: string;
  'ui:autocomplete'?: string;
  default?: string | number | boolean;
  items?: { [key: string]: any };
  'ui:*items'?: { [key: string]: any };
  multipleOf?: number | null;
  minimum?: number | null;
  exclusiveMinimum?: number | null;
  maximum?: number | null;
  exclusiveMaximum?: number | null;
  enum?: (number | string)[];
  enumNames?: string[] | null;
  description?: string;
}